refactor: del canvas-visualiseringene på delte primitiver
Flytter useVisualizationEnvironment til components/visualization og legger til
canvas.ts (createCssVarReader, syncCanvasResolution, observeThemeChange).
AStar og AutoSnakeBackground bruker nå disse i stedet for duplisert kode, og
snake-logikken er delt i snake/{settings,game,colors}.ts.
